Move default map type to ext
