+++ ground/src/main/java/com/google/android/ground/exts/GmsExt.kt
@@ -1,5 +1,6 @@
 package com.google.android.ground.exts
 
+import com.google.android.gms.maps.GoogleMap
 import com.google.android.ground.model.geometry.Coordinate
 import com.google.android.ground.model.geometry.Geometry
 import com.google.android.ground.ui.map.Bounds
@@ -8,7 +9,7 @@ import com.google.android.ground.ui.map.gms.toLatLng
 import com.google.android.ground.ui.map.gms.toModelObject
 
 /** Extensions for indirectly using GMS functions in map-provider agnostic codebase. */
-object GeometryExt {
+object GmsExt {
 
   fun Bounds.contains(coordinate: Coordinate): Boolean {
     return toGoogleMapsObject().contains(coordinate.toGoogleMapsObject())
@@ -22,4 +23,6 @@ object GeometryExt {
   fun Bounds.center(): Coordinate {
     return toGoogleMapsObject().center.toModelObject()
   }
+
+  fun defaultMapType(): Int = GoogleMap.MAP_TYPE_HYBRID
 }
